Expand bridging_roles API.
This adds channel-based functions that already existed
for bridge_channels. The reason for this is that my attended
transfer push function needs to check for the existence of roles
on a channel. I can't use the bridge channel at that point because
bridging roles are not established until after the push callback
is called.