Fix up some referencing issues.
There was a refleak in my move and merge functions since bridge channels
were not being unreffed.

There was also a misunderstanding about ast_bridge_impart's practice of
gobbling the reference of the channel that is being added. This led to
an overabundance of unrefs for the transfer target.

There appears to be a refleak of some sort still, since "core stop now"
after performing a transfer results in a delay.