Adjust the state machine to make more sense with the new superstates.
Once in the recalling superstate, there is no such thing as a transferer
or transfer target. Intead, the party being recalled is always the recall
target, no matter the party's original role in the transfer. I've created
two new stimuli for this and gotten rid of the transferer answer stimulus
since that no longer makes sense.

Testing shows things to still work correctly.