Handle a previously unhandled case.
If the transferee hangs up during the Consulting state, then it doesn't
make sense to just destroy everything. Instead, we play an error sound
to the transferer and change the bridge that the transferer and transfer
target are in back to normal. They can then continue on as they please.