OneTouchRecord: Make so Monitor/MixMonitor can be toggled/started/stopped.
The OneTouchRecord start and stop option will only start or stop the
recording.  If OneTouchRecord is already doing what is requested then only
the invoker hears the courtesy tone and/or start/stop recording message.

The new feature is written so it we could easily add explicit start/stop
recording DTMF hooks for Monitor and MixMonitor.

* Made agents.conf recordagentcalls option do start only OneTouchRecord
when the agent initially connects to the caller.