Refactor mixmonitor setup on queues.
The BUGBUG comment suggested monitoring the bridge instead of setting up
a mixmonitor on the calling channel. For now anyway, I have not done this.
However, having mixmonitor setup as a single routine makes it much simpler
to do this if desired. However, I think it's debatable whether we actually
want to be monitoring the bridge instead of the caller.

In addition to the refactoring, I actually fixed a bug. The
MONITOR_EXEC channel variable went through its escaping and variable
substitution routine, but the unescaped variable name was being passed
to the MixMonitor application rather than the escaped name. I did not
find any open issues pertaining to this.