Completely rework local optimization handling.
The issue here is that when local channels are involved, the other
side of the local channel is a complete mystery to you. There could
be one or more parties on the other side of the local channel. If
more than one party is in that local channel, then the result could
be that we end up with the queue needing to monitor a multi-party
call. Any channels that originate from the caller's side of the
original bridge will be considered callers and any channels that
originate from the agent's side of the of the bridge will be
considered agents. We keep a container for each and modify the
containers as necessary when local channel optimizations occur.
We also track hangups accordingly and can see when the last
of the callers or last of the agents has hung up. It is not
until one of these happens that we send an agent complete event.

This is pretty overkill considering that 99.999% of the time,
even if there are local channels involved, we'll be dealing with
a two-party call. However, for those weirdos out there that end
up with local channels that call local channels that call into
a bridge that then splits out into seven local channels that
each bridge into more local channels that then bridge to actual
real things, this should work for them too. Hopefully.

Oh, this isn't tested. Testing and bugfixing begins Monday.