Blocked revisions 395020
........
Add an upgrade note for libuuid dependency; remove note in CHANGES

This patch notes that libuuid is now a dependency for res_rtp_asterisk; this
was introduced in between 11.4.0 and 11.5.0 to resolve a dependency for
pjproject, which res_rtp_asterisk uses for ICE/STUN/TURN support.

It also removes a conflicting note from CHANGES. While support for playing
prompts to the first participant was added for app_queue, it was disabled
by default and an option added to enable it. That was properly noted in the
UPGRADE.txt file.
........

Merged revisions 395021 from file:///srv/subversion/repos/asterisk/trunk
........

Merged revisions 395028 from file:///srv/subversion/repos/asterisk/team/rmudgett/bridge_phase