Resolve a race condition during local channel optimizations.
I need to get the channels that are in the involved bridges of
a local channel optimization. My previous code got a channel from
a channel snapshot and then got that channel's bridge. The problem
is that the channel may already be out of the bridge by the time
we process the local channel optimization. Instead, we just grab
the latest bridge snapshot for the channel snapshot we have. This
is more tolerant of the channel having been removed from the
bridging system. This also makes the code likely to be more efficient
since there is less locking, and we are not having to look into
the internals of an actual bridge, but rather a snapshot.