Create a queue topic on which to publish agent complete messages.
Most stasis messages in app_queue are published on the caller's channel
topic. This works fine since the messages are published prior to bridging.
The problem is that once the caller and member are bridged, the caller's
channel may no longer be the same. It may be that the caller's channel
gets optimized away and a new channel becomes the caller's channel. As such,
publishing on this new caller's topic may be a bit odd. Instead, we now publish
the completion event on the queue's topic.

It may be worthwhile to publish other messages on the queue's topic but that
can be a different task.